requirements.txt
discord.py
flask
schedule
asyncio
requests
twstock
lxml
bs4
opencc
python-dotenv